requirements.txt
arxiv
pymupdf
python-dotenv
chromadb>=0.5.0
pydantic-settings
sentence-transformers
rank-bm25
llama-index
llama-index-readers-file
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
llama-index-llms-ollama
